Switch to reparametrizations and refactor
